move write/read logic from core to content.py
